requirements.txt
#
# Clean from-scratch (tested on MacPro/M1 - Dec 2021)
#
newspaper3k
transformers
spacy
bert-extractive-summarizer
summarizer
torch
sentencepiece  # note: requires installation - this is just a py wrapper


# bert-extractive-summarizer==0.8.1
# blis==0.7.5
# catalogue==2.0.6
# certifi==2021.10.8
# charset-normalizer==2.0.7
# click==8.0.3
# cymem==2.0.6
# filelock==3.3.2
# huggingface-hub==0.1.2
# idna==3.3
# Jinja2==3.0.3
# joblib==1.1.0
# langcodes==3.3.0
# MarkupSafe==2.0.1
# murmurhash==1.0.6
# newspaper3k==0.2.8
# nltk==3.6.5
# numpy==1.21.4
# packaging==21.2
# pathy==0.6.1
# preshed==3.0.6
# pydantic==1.8.2
# pyparsing==2.4.7
# PyYAML==6.0
# regex==2021.11.10
# requests==2.26.0
# sacremoses==0.0.46
# scikit-learn==1.0.1
# scipy==1.7.2
# sentencepiece==0.1.96
# six==1.16.0
# smart-open==5.2.1
# spacy==3.2.0
# spacy-legacy==3.0.8
# spacy-loggers==1.0.1
# srsly==2.4.2
# summarizer==0.0.7
# thinc==8.0.13
# threadpoolctl==3.0.0
# tokenizers==0.10.3
# torch==1.10.0
# tqdm==4.62.3
# transformers==4.12.3
# typer==0.4.0
# typing-extensions==3.10.0.2
# urllib3==1.26.7
# wasabi==0.8.2
